+++ doc/whatsnew.rst
@@ -11,6 +11,7 @@ What's new?
 Next release
 ============
 
+- Add data cleaning for the T004 historical input data source (:pull:`59`).
 - Update :doc:`usage` documentation (:issue:`41`, :pull:`43`).
 - Increase minimum Python version to 3.7 and ensure compatibility (:issue:`41`, :pull:`43`).
 - Add the :doc:`glossary` page (:pull:`42`).